Call PhoenixWorker as a local_worker to better work with topology
Tree mode requires a distant_worker to connect between the root and the
gateways. Phoenix can't do this, so allow ssh to remain the
distant_worker and call task.shell() with remote=False to force the use
of a local_worker as the last step.

Note: ClusterShell hardcodes the local_worker as exec in tree mode. This
requires patches to ClusterShell until I can get it upstreamed.